sqs-reader/requirements.txt
boto3==1.18.16
botocore==1.21.16
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
jmespath==0.10.0
python-dateutil==2.8.2
python-dotenv==0.19.0
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.6
